Refactor Dashboard to allow for wallet service to be passed
